requirements.txt
fastmcp
requests
requests-negotiate-sspi
truststore
python-dotenv
xmltodict
